Extract Page::Sitepressed
